requirements.txt
BeautifulSoup==3.2.1
Django==1.5.1
Django-tinymce-filebrowser==0.1.1
Mako==0.7.3
Markdown==2.3.1
MarkupSafe==0.15
PAM==0.4.2
Pillow==2.0.0
Pygments==1.6
South==0.8.2
Twisted-Core==12.3.0
Twisted-Names==12.3.0
Twisted-Web==12.3.0
apt-xapian-index==0.45
aptoncd==0.1.98-bzr117-1.2
apturl==0.4.1ubuntu4
argparse==1.2.1
chardet==2.0.1
configglue==1.0
debtagshw==0.1
defer==1.0.6
dirspec==4.2.0
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-ckeditor==4.0.2
django-pygments==0.1
django-tinymce==1.5.1
django-toolbelt==0.0.1
duplicity==0.6.21
feedparser==5.1.2
gunicorn==18.0
httplib2==0.7.7
lxml==3.1.0
oauthlib==0.3.7
oneconf==0.3.3
paramiko==1.7.7.1
pexpect==2.4
piston-mini-client==0.7.5
protobuf==2.4.1
psycopg2==2.4.5
pyOpenSSL==0.13
pycrypto==2.6
pycups==1.9.62
pycurl==7.19.0
pygobject==3.8.0
pygpgme==0.3
pyinotify==0.9.3
pyserial==2.6
pysmbc==1.0.13
python-apt==0.8.8ubuntu6
python-debian==0.1.21-nmu2ubuntu1
pyxdg==0.25
reportlab==2.6
sessioninstaller==0.0.0
six==1.2.0
sorl-thumbnail==11.12
static==0.4
system-service==0.1.6
uTidylib==0.2
virtualenv==1.10.1
wsgiref==0.1.2
zope.interface==4.0.5
